+++ app/Http/Livewire/Series/ShowSeries.php
@@ -126,7 +126,7 @@ class ShowSeries extends Component
         toastr()->livewire()->addSuccess(__(':name has been updated', ['name' => $volume->series->name . ' ' . $volume->number]));
     }
 
-    public function update(SeriesService $seriesService): void
+    public function update(SeriesService $seriesService)
     {
         try {
             $this->series = $seriesService->refreshMetadata($this->series);
@@ -141,7 +141,9 @@ class ShowSeries extends Component
 
             $seriesService->updateVolumes($this->series);
 
-            toastr()->livewire()->addSuccess(__(':name has been updated', ['name' => $this->series->name]));
+            toastr()->addSuccess(__(':name has been updated', ['name' => $this->series->name]));
+
+            return redirect(request()->header('Referer'));
         } catch (Exception $exception) {
             Log::error('Error while updating series via API', ['exception' => $exception]);
             toastr()->livewire()->addError(__(':name could not be updated', ['name' => $this->series->name]));