+++ resources/views/livewire/volumes/edit-volume.blade.php
@@ -84,6 +84,17 @@
                             @enderror
                         </div>
                     </div>
+                    <div class="row mt-1">
+                        <div class="col-md-12">
+                            <label for="volume.image_url" class="col-form-label">{{ __('Image URL') }}</label>
+                            <input id="volume.image_url" name="volume.image_url" type="text" class="form-control @error('volume.image_url') is-invalid @enderror" wire:model='volume.image_url'>
+                            @error('volume.image_url')
+                                <span class="invalid-feedback" role="alert">
+                                    <strong>{{ $message }}</strong>
+                                </span>
+                            @enderror
+                        </div>
+                    </div>
                     <div class="row mt-3">
                         <div class="col-md-12">
                             <div class="form-check">